+++ webviz_config/__init__.py
@@ -11,6 +11,7 @@ from ._localhost_token import LocalhostToken
 from ._is_reload_process import is_reload_process
 from ._plugin_abc import WebvizPluginABC, EncodedFile, ZipFileMember
 from ._shared_settings_subscriptions import SHARED_SETTINGS_SUBSCRIPTIONS
+from ._oauth2 import Oauth2
 
 try:
     __version__ = version("webviz-config")